requirements.txt
Django
pillow
pdfminer.six
PyMuPDF
pytesseract
django-crispy-forms
reportlab
PyPDF2